requirements.txt
selenium==3.141.0
python-dotenv==0.17.1
requests==2.25.1
